New version of clippy has a lot of new lints
Lots of new clippy lints due to toolchain upgrade.

Made the Message::error field pub(crate) again, it was the odd one out
and it seemed a reasonable way to shut up clippy.

+++ src/test_utils.rs
@@ -504,7 +504,7 @@ async fn log_msg(context: &Context, prefix: impl AsRef<str>, msg: &Message) {
         &contact_name,
         contact_id,
         msgtext.unwrap_or_default(),
-        if msg.get_from_id() == 1 as libc::c_uint {
+        if msg.get_from_id() == 1u32 {
             ""
         } else if msg.get_state() == MessageState::InSeen {
             "[SEEN]"
